test: add integration test protocol (prepare for swift testing :))
