+++ Tests/IntegrationTests/BugReportTests.swift
@@ -16,23 +16,7 @@ import NIOPosix
 import OracleNIO
 import XCTest
 
-final class BugReportTests: XCTestCase {
-
-    private var group: EventLoopGroup!
-
-    private var eventLoop: EventLoop { self.group.next() }
-
-    override func setUpWithError() throws {
-        try super.setUpWithError()
-        XCTAssertTrue(isLoggingConfigured)
-        self.group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
-    }
-
-    override func tearDownWithError() throws {
-        try self.group?.syncShutdownGracefully()
-        self.group = nil
-        try super.tearDownWithError()
-    }
+final class BugReportTests: XCTIntegrationTest {
 
     func testRowsFetchFailsWithDecodingError() async {
         do {
@@ -54,8 +38,6 @@ final class BugReportTests: XCTestCase {
                     readonly number
                 )
                 """
-            let connection = try await OracleConnection.test(on: self.eventLoop)
-            defer { XCTAssertNoThrow(try connection.syncClose()) }
             _ = try? await connection.query("DROP TABLE EXPORT_TABLE")
             try await connection.query(schema)
             try await connection.query(
