+++ Tests/IntegrationTests/IntegrationTest.swift
@@ -1,0 +1,5 @@
+import OracleNIO
+
+protocol IntegrationTest {
+    var connection: OracleConnection! { get set }
+}
